+++ lib/asin/client.rb
@@ -3,7 +3,7 @@ require 'rexml/document' # https://github.com/phoet/asin/pull/23
 require 'crack/xml'
 require 'cgi'
 require 'base64'
-require 'rash'
+require 'sc_hash'
 
 module ASIN
   module Client
@@ -216,7 +216,7 @@ module ASIN
     end
 
     def handle_type(data, type)
-      Hashie::Rash.new(data)
+      Hashie::SCHash.new(data)
     end
 
     def create_item_params(items)